backend/functions/requirements.txt
firebase_functions
stripe
markdown2
google-cloud-pubsub
